homework10-13-FastAPI+SQLAlchemy+Alembic+Auth+Redis+Tests/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.0
requests==2.31.0
pytest-cov==4.1.0